Fix pgbouncer race condition, slightly clean code
There is a problem with PGBouncer where occasionally, the RELOAD command will be called too quickly after changing the symlink, and pgbouncer will pull the old config file, failing to swap to the new master.  So, we now perform helath checks on pgbouncer itself, and when the master we're supposed to be using is live but the pgbouncer is not, we issue a new RELOAD command.  In order to ensure that Twitch's (very tight) recovery window is respected, we also put rebouncer into "aggressive mode" after a failover, checking every 100ms instead of the configured check window.  This should allow rebouncer to complete the failover within 100ms even when the initial RELOAD fails.

This added some complexity to the code and things started to unravel so I also performed some basic refactoring, moving the server concept (representing both the master/hotspares themselves but also the rebouncer) to a separate file.